allocators: Apply Allocator named requirements to secure_allocator::deallocate
The named requirements for Allocator are here:
    https://en.cppreference.com/w/cpp/named_req/Allocator

It gives the following requirements for `a.deallocate(p, n)`:

> Deallocates storage pointed to `p`, which must be a value returned by
> a previous call to `allocate` that has not been invalidated by an
> intervening call to `deallocate`. `n` must match the value previously
> passed to `allocate`. Does not throw exceptions.

This explicitly excludes both calling `secure_allocator::deallocate` on
`nullptr`, and throwing exceptions of any kind. We address the former by
asserting that `p != nullptr` (because `LockedPool::free` does support
`nullptr`), and we address the latter by adding the `noexcept` keyword
(which the example C++11 allocator on the page above also uses).