Allow vdxfuniobjects to be arrays
When deserializing a stream into uniobjects if there is more than one it will put them into an array.  however on encoding to a vector this does not happen.  So the only way to put uniobjects in of the same time is create an uniobject like this with the same key names.

```js
{
   "objectdata":{
      "i4GC1YGEVD21afWudGoFJVdnfjJ5XWnCQv":{
         "version":1,
         "mimetype":"text/plain",
         "objectdata":{
            "message":"IBM"
         },
         "label":"Employee"
      },
      "i4GC1YGEVD21afWudGoFJVdnfjJ5XWnCQv":{
         "version":1,
         "mimetype":"text/plain",
         "objectdata":{
            "message":"2000 - 2002"
         },
         "label":"date"
      }
   }
}